Types cleanup.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@17852 27425a3e-05d8-49a3-a47f-9c15f0e5edd8

+++ os/sb/host/sbposix.c
@@ -65,19 +65,21 @@ static msg_t create_descriptor(sb_ioblock_t *iop,
   return CH_RET_EMFILE;
 }
 
-static int sb_io_stat(sb_class_t *sbp, const char *path, struct stat *statbuf) {
+static uint32_t sb_io_stat(sb_class_t *sbp,
+                           const char *path,
+                           struct stat *statbuf) {
   msg_t ret;
   vfs_stat_t vstat;
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   if (!sb_is_valid_write_range(sbp, (void *)statbuf, sizeof (struct stat))) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  ret = (int)vfsDrvStat(sbp->io.vfs_driver, path, &vstat);
+  ret = vfsDrvStat(sbp->io.vfs_driver, path, &vstat);
   if (!CH_RET_IS_ERROR(ret)) {
     memset((void *)statbuf, 0, sizeof (struct stat));
     statbuf->st_mode  = (mode_t)vstat.mode;
@@ -86,15 +88,15 @@ static int sb_io_stat(sb_class_t *sbp, const char *path, struct stat *statbuf) {
     /* TODO st_blocks, st_blksize, st_ino, timespecs.*/
   }
 
-  return ret;
+  return (uint32_t)ret;
 }
 
-static int sb_io_open(sb_class_t *sbp, const char *path, int flags) {
+static uint32_t sb_io_open(sb_class_t *sbp, const char *path, int flags) {
   vfs_node_c *np = NULL;
   msg_t ret;
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   do {
@@ -104,34 +106,34 @@ static int sb_io_open(sb_class_t *sbp, const char *path, int flags) {
     ret = create_descriptor(&sbp->io, np);
     CH_BREAK_ON_ERROR(ret);
 
-    return (int)ret;
+    return (uint32_t)ret;
   } while (false);
 
   if (np != NULL) {
     vfsClose(np);
   }
 
-  return (int)ret;
+  return (uint32_t)ret;
 }
 
-static int sb_io_close(sb_class_t *sbp, int fd) {
+static uint32_t sb_io_close(sb_class_t *sbp, int fd) {
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   vfsClose(sbp->io.vfs_nodes[fd]);
   sbp->io.vfs_nodes[fd] = NULL;
 
-  return CH_RET_SUCCESS;
+  return (uint32_t)CH_RET_SUCCESS;
 }
 
-static int sb_io_dup(sb_class_t *sbp, int fd) {
+static uint32_t sb_io_dup(sb_class_t *sbp, int fd) {
   vfs_node_c *np;
   msg_t ret;
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   /* Node associated to the existing file descriptor.*/
@@ -145,21 +147,21 @@ static int sb_io_dup(sb_class_t *sbp, int fd) {
     vfsClose(np);
   }
 
-  return (int)ret;
+  return (uint32_t)ret;
 }
 
-static int sb_io_dup2(sb_class_t *sbp, int oldfd, int newfd) {
+static uint32_t sb_io_dup2(sb_class_t *sbp, int oldfd, int newfd) {
 
   if (!sb_is_existing_descriptor(&sbp->io, oldfd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (!sb_is_valid_descriptor(newfd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (oldfd == newfd) {
-    return (int)newfd;
+    return (uint32_t)newfd;
   }
 
   if (sbp->io.vfs_nodes[newfd] != NULL) {
@@ -168,19 +170,19 @@ static int sb_io_dup2(sb_class_t *sbp, int oldfd, int newfd) {
 
   sbp->io.vfs_nodes[newfd] = (vfs_node_c *)roAddRef(sbp->io.vfs_nodes[oldfd]);
 
-  return (int)newfd;
+  return (uint32_t)newfd;
 }
 
-static int sb_io_fstat(sb_class_t *sbp, int fd, struct stat *statbuf) {
+static uint32_t sb_io_fstat(sb_class_t *sbp, int fd, struct stat *statbuf) {
   msg_t ret;
   vfs_stat_t vstat;
 
   if (!sb_is_valid_write_range(sbp, (void *)statbuf, sizeof (struct stat))) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   ret = vfsGetNodeStat(sbp->io.vfs_nodes[fd], &vstat);
@@ -191,81 +193,88 @@ static int sb_io_fstat(sb_class_t *sbp, int fd, struct stat *statbuf) {
     statbuf->st_nlink = 1;
   }
 
-  return ret;
+  return (uint32_t)ret;
 }
 
-static ssize_t sb_io_read(sb_class_t *sbp, int fd, void *buf, size_t count) {
+static uint32_t sb_io_read(sb_class_t *sbp, int fd, void *buf, size_t count) {
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (VFS_MODE_S_ISDIR(sbp->io.vfs_nodes[fd]->mode)) {
-    return CH_RET_EISDIR;
+    return (uint32_t)CH_RET_EISDIR;
   }
 
   if (count == (size_t)0) {
-    return 0;
+    return (uint32_t)0;
   }
 
   if (!sb_is_valid_write_range(sbp, buf, count)) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return vfsReadFile((vfs_file_node_c *)sbp->io.vfs_nodes[fd], buf, count);
+  return (uint32_t)vfsReadFile((vfs_file_node_c *)sbp->io.vfs_nodes[fd],
+                               buf,
+                               count);
 }
 
-static ssize_t sb_io_write(sb_class_t *sbp, int fd, const void *buf, size_t count) {
+static uint32_t sb_io_write(sb_class_t *sbp,
+                            int fd,
+                            const void *buf,
+                            size_t count) {
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (VFS_MODE_S_ISDIR(sbp->io.vfs_nodes[fd]->mode)) {
-    return CH_RET_EISDIR;
+    return (uint32_t)CH_RET_EISDIR;
   }
 
   if (count == (size_t)0) {
-    return 0;
+    return (uint32_t)0;
   }
 
   if (!sb_is_valid_read_range(sbp, buf, count)) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return vfsWriteFile((vfs_file_node_c *)sbp->io.vfs_nodes[fd], buf, count);
+  return (uint32_t)vfsWriteFile((vfs_file_node_c *)sbp->io.vfs_nodes[fd],
+                                buf,
+                                count);
 }
 
-static off_t sb_io_lseek(sb_class_t *sbp, int fd, off_t offset, int whence) {
+static uint32_t sb_io_lseek(sb_class_t *sbp, int fd, off_t offset, int whence) {
   vfs_offset_t pos;
 
   if ((whence != SEEK_SET) && (whence != SEEK_CUR) && (whence != SEEK_END)) {
-    return CH_RET_EINVAL;
+    return (uint32_t)CH_RET_EINVAL;
   }
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (VFS_MODE_S_ISDIR(sbp->io.vfs_nodes[fd]->mode)) {
-    return CH_RET_EISDIR;
+    return (uint32_t)CH_RET_EISDIR;
   }
 
   if (!VFS_MODE_S_ISREG(sbp->io.vfs_nodes[fd]->mode)) {
-    return CH_RET_ESPIPE;
+    return (uint32_t)CH_RET_ESPIPE;
   }
 
   pos = vfsSetFilePosition((struct vfs_file_node *)sbp->io.vfs_nodes[fd],
                            offset,
                            whence);
   if (CH_RET_IS_ERROR(pos)) {
-    return (off_t)pos;
+    return (uint32_t)pos;
   }
 
-  return (off_t)vfsGetFilePosition((struct vfs_file_node *)sbp->io.vfs_nodes[fd]);
+  return (uint32_t)vfsGetFilePosition((struct vfs_file_node *)sbp->io.vfs_nodes[fd]);
 }
 
-static ssize_t sb_io_getdents(sb_class_t *sbp, int fd, void *buf, size_t count) {
+static uint32_t sb_io_getdents(sb_class_t *sbp, int fd, void *buf, size_t count) {
   vfs_shared_buffer_t *shbuf;
   vfs_direntry_info_t *dip;
   msg_t ret;
@@ -273,24 +282,24 @@ static ssize_t sb_io_getdents(sb_class_t *sbp, int fd, void *buf, size_t count) 
   size_t max_entry;
 
   if (count == (size_t)0) {
-    return (ssize_t)CH_RET_EINVAL;
+    return (uint32_t)CH_RET_EINVAL;
   }
 
   if (!sb_is_valid_write_range(sbp, buf, count)) {
-    return (ssize_t)CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   min_entry = sizeof (struct dirent) + (size_t)1;
   if (count < min_entry) {
-    return (ssize_t)CH_RET_EINVAL;
+    return (uint32_t)CH_RET_EINVAL;
   }
 
   if (!sb_is_existing_descriptor(&sbp->io, fd)) {
-    return (ssize_t)CH_RET_EBADF;
+    return (uint32_t)CH_RET_EBADF;
   }
 
   if (!VFS_MODE_S_ISDIR(sbp->io.vfs_nodes[fd]->mode)) {
-    return (ssize_t)CH_RET_ENOTDIR;
+    return (uint32_t)CH_RET_ENOTDIR;
   }
 
   max_entry = sizeof (struct dirent) + (size_t)VFS_CFG_NAMELEN_MAX + (size_t)1;
@@ -347,67 +356,69 @@ static ssize_t sb_io_getdents(sb_class_t *sbp, int fd, void *buf, size_t count) 
 
   vfs_buffer_release(shbuf);
 
-  return (ssize_t)ret;
+  return (uint32_t)ret;
 }
 
-static int sb_io_chdir(sb_class_t *sbp, const char *path) {
+static uint32_t sb_io_chdir(sb_class_t *sbp, const char *path) {
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return (int)vfsDrvChangeCurrentDirectory(sbp->io.vfs_driver, path);
+  return (uint32_t)vfsDrvChangeCurrentDirectory(sbp->io.vfs_driver, path);
 }
 
-static int sb_io_getcwd(sb_class_t *sbp, char *buf, size_t size) {
+static uint32_t sb_io_getcwd(sb_class_t *sbp, char *buf, size_t size) {
 
   if (!sb_is_valid_write_range(sbp, buf, size)) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   /* Note, it does not return a pointer to the buffer as required by Posix,
      this has to be handled on the user-side library.*/
-  return vfsDrvGetCurrentDirectory(sbp->io.vfs_driver, buf, size);
+  return (uint32_t)vfsDrvGetCurrentDirectory(sbp->io.vfs_driver, buf, size);
 }
 
-static int sb_io_unlink(sb_class_t *sbp, const char *path) {
+static uint32_t sb_io_unlink(sb_class_t *sbp, const char *path) {
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return (int)vfsDrvUnlink(sbp->io.vfs_driver, path);
+  return (uint32_t)vfsDrvUnlink(sbp->io.vfs_driver, path);
 }
 
-static int sb_io_rename(sb_class_t *sbp, const char *oldpath, const char *newpath) {
+static uint32_t sb_io_rename(sb_class_t *sbp,
+                             const char *oldpath,
+                             const char *newpath) {
 
   if (sb_check_string(sbp, (void *)oldpath, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
   if (sb_check_string(sbp, (void *)newpath, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return (int)vfsDrvRename(sbp->io.vfs_driver, oldpath, newpath);
+  return (uint32_t)vfsDrvRename(sbp->io.vfs_driver, oldpath, newpath);
 }
 
-static int sb_io_mkdir(sb_class_t *sbp, const char *path, mode_t mode) {
+static uint32_t sb_io_mkdir(sb_class_t *sbp, const char *path, mode_t mode) {
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return (int)vfsDrvMkdir(sbp->io.vfs_driver, path, (vfs_mode_t)mode);
+  return (uint32_t)vfsDrvMkdir(sbp->io.vfs_driver, path, (vfs_mode_t)mode);
 }
 
-static int sb_io_rmdir(sb_class_t *sbp, const char *path) {
+static uint32_t sb_io_rmdir(sb_class_t *sbp, const char *path) {
 
   if (sb_check_string(sbp, (void *)path, VFS_CFG_PATHLEN_MAX + 1) == (size_t)0) {
-    return CH_RET_EFAULT;
+    return (uint32_t)CH_RET_EFAULT;
   }
 
-  return (int)vfsDrvRmdir(sbp->io.vfs_driver, path);
+  return (uint32_t)vfsDrvRmdir(sbp->io.vfs_driver, path);
 }
 
 /*===========================================================================*/
@@ -437,78 +448,70 @@ void sb_sysc_stdio(sb_class_t *sbp, struct port_extctx *ectxp) {
 
   switch (ectxp->r0) {
   case SB_POSIX_OPEN:
-    ectxp->r0 = (uint32_t)sb_io_open(sbp,
-                                     (const char *)ectxp->r1,
-                                     (int)ectxp->r2);
+    ectxp->r0 = sb_io_open(sbp, (const char *)ectxp->r1, (int)ectxp->r2);
     break;
   case SB_POSIX_CLOSE:
-    ectxp->r0 = (uint32_t)sb_io_close(sbp, (int)ectxp->r1);
+    ectxp->r0 = sb_io_close(sbp, (int)ectxp->r1);
     break;
   case SB_POSIX_DUP:
-    ectxp->r0 = (uint32_t)sb_io_dup(sbp, (int)ectxp->r1);
+    ectxp->r0 = sb_io_dup(sbp, (int)ectxp->r1);
     break;
   case SB_POSIX_DUP2:
-    ectxp->r0 = (uint32_t)sb_io_dup2(sbp,
-                                     (int)ectxp->r1,
-                                     (int)ectxp->r2);
+    ectxp->r0 = sb_io_dup2(sbp, (int)ectxp->r1, (int)ectxp->r2);
     break;
   case SB_POSIX_FSTAT:
-    ectxp->r0 = (uint32_t)sb_io_fstat(sbp,
-                                      (int)ectxp->r1,
-                                      (struct stat *)ectxp->r2);
+    ectxp->r0 = sb_io_fstat(sbp, (int)ectxp->r1, (struct stat *)ectxp->r2);
     break;
   case SB_POSIX_READ:
-    ectxp->r0 = (uint32_t)sb_io_read(sbp,
-                                     (int)ectxp->r1,
-                                     (void *)ectxp->r2,
-                                     (size_t)ectxp->r3);
+    ectxp->r0 = sb_io_read(sbp,
+                           (int)ectxp->r1,
+                           (void *)ectxp->r2,
+                           (size_t)ectxp->r3);
     break;
   case SB_POSIX_WRITE:
-    ectxp->r0 = (uint32_t)sb_io_write(sbp,
-                                      (int)ectxp->r1,
-                                      (const void *)ectxp->r2,
-                                      (size_t)ectxp->r3);
+    ectxp->r0 = sb_io_write(sbp,
+                            (int)ectxp->r1,
+                            (const void *)ectxp->r2,
+                            (size_t)ectxp->r3);
     break;
   case SB_POSIX_LSEEK:
-    ectxp->r0 = (uint32_t)sb_io_lseek(sbp,
-                                      (int)ectxp->r1,
-                                      (off_t)ectxp->r2,
-                                      (int)ectxp->r3);
+    ectxp->r0 = sb_io_lseek(sbp,
+                            (int)ectxp->r1,
+                            (off_t)ectxp->r2,
+                            (int)ectxp->r3);
     break;
   case SB_POSIX_GETDENTS:
-    ectxp->r0 = (uint32_t)sb_io_getdents(sbp,
-                                         (int)ectxp->r1,
-                                         (void *)ectxp->r2,
-                                         (size_t)ectxp->r3);
+    ectxp->r0 = sb_io_getdents(sbp,
+                               (int)ectxp->r1,
+                               (void *)ectxp->r2,
+                               (size_t)ectxp->r3);
     break;
   case SB_POSIX_CHDIR:
-    ectxp->r0 = (uint32_t)sb_io_chdir(sbp, (const char *)ectxp->r1);
+    ectxp->r0 = sb_io_chdir(sbp, (const char *)ectxp->r1);
     break;
   case SB_POSIX_GETCWD:
-    ectxp->r0 = (uint32_t)sb_io_getcwd(sbp,
-                                       (char *)ectxp->r1,
-                                       (size_t)ectxp->r2);
+    ectxp->r0 = sb_io_getcwd(sbp, (char *)ectxp->r1, (size_t)ectxp->r2);
     break;
   case SB_POSIX_UNLINK:
-    ectxp->r0 = (uint32_t)sb_io_unlink(sbp, (const char *)ectxp->r1);
+    ectxp->r0 = sb_io_unlink(sbp, (const char *)ectxp->r1);
     break;
   case SB_POSIX_RENAME:
-    ectxp->r0 = (uint32_t)sb_io_rename(sbp,
-                                       (const char *)ectxp->r1,
-                                       (const char *)ectxp->r2);
+    ectxp->r0 = sb_io_rename(sbp,
+                             (const char *)ectxp->r1,
+                             (const char *)ectxp->r2);
     break;
   case SB_POSIX_MKDIR:
-    ectxp->r0 = (uint32_t)sb_io_mkdir(sbp,
-                                      (const char *)ectxp->r1,
-                                      (mode_t)ectxp->r2);
+    ectxp->r0 = sb_io_mkdir(sbp,
+                            (const char *)ectxp->r1,
+                            (mode_t)ectxp->r2);
     break;
   case SB_POSIX_RMDIR:
-    ectxp->r0 = (uint32_t)sb_io_rmdir(sbp, (const char *)ectxp->r1);
+    ectxp->r0 = sb_io_rmdir(sbp, (const char *)ectxp->r1);
     break;
   case SB_POSIX_STAT:
-    ectxp->r0 = (uint32_t)sb_io_stat(sbp,
-                                     (const char *)ectxp->r1,
-                                     (struct stat *)ectxp->r2);
+    ectxp->r0 = sb_io_stat(sbp,
+                           (const char *)ectxp->r1,
+                           (struct stat *)ectxp->r2);
     break;
   default:
     ectxp->r0 = (uint32_t)CH_RET_ENOSYS;
